kafka17109: lock backoff-retry
