+++ streams/src/main/java/org/apache/kafka/streams/processor/internals/TaskManager.java
@@ -28,6 +28,7 @@ import org.apache.kafka.common.Metric;
 import org.apache.kafka.common.MetricName;
 import org.apache.kafka.common.TopicPartition;
 import org.apache.kafka.common.errors.TimeoutException;
+import org.apache.kafka.common.utils.ExponentialBackoff;
 import org.apache.kafka.common.utils.LogContext;
 import org.apache.kafka.common.utils.Time;
 import org.apache.kafka.streams.errors.LockException;
@@ -103,6 +104,8 @@ public class TaskManager {
 
     // includes assigned & initialized tasks and unassigned tasks we locked temporarily during rebalance
     private final Set<TaskId> lockedTaskDirectories = new HashSet<>();
+
+    private final HashMap<TaskId, BackoffRecord> taskIdToBackoffRecord = new HashMap<>();
 
     private final ActiveTaskCreator activeTaskCreator;
     private final StandbyTaskCreator standbyTaskCreator;
@@ -1007,14 +1010,23 @@ public class TaskManager {
     }
 
     private void addTaskToStateUpdater(final Task task) {
+        final long nowMs = System.currentTimeMillis();
         try {
-            task.initializeIfNeeded();
-            stateUpdater.add(task);
+            if (canTryLock(task.id(), nowMs)) {
+                task.initializeIfNeeded();
+                stateUpdater.add(task);
+                taskIdToBackoffRecord.remove(task.id());
+            } else {
+                final String errorMsg = String.format("Task " + task.id() + " is still not allowed to retry acquiring the state directory lock");
+                log.trace(errorMsg);
+                handleUnsuccessfulLockAcquiring(task, nowMs);
+                throw new RuntimeException(errorMsg);
+            }
         } catch (final LockException lockException) {
             // The state directory may still be locked by another thread, when the rebalance just happened.
             // Retry in the next iteration.
             log.info("Encountered lock exception. Reattempting locking the state in the next iteration.", lockException);
-            tasks.addPendingTasksToInit(Collections.singleton(task));
+            handleUnsuccessfulLockAcquiring(task, nowMs);
         }
     }
 
@@ -2116,5 +2128,43 @@ public class TaskManager {
     // for testing only
     void addTask(final Task task) {
         tasks.addTask(task);
+    }
+
+    private void handleUnsuccessfulLockAcquiring(final Task task, final long nowMs) {
+        updateOrCreateBackoffRecord(task.id(), nowMs);
+        tasks.addPendingTasksToInit(Collections.singleton(task));
+    }
+
+    private boolean canTryLock(final TaskId taskId, final long nowMs) {
+        return !taskIdToBackoffRecord.containsKey(taskId) || taskIdToBackoffRecord.get(taskId).canAttempt(nowMs);
+    }
+
+    private void updateOrCreateBackoffRecord(final TaskId taskId, final long nowMs) {
+        if (taskIdToBackoffRecord.containsKey(taskId)) {
+            taskIdToBackoffRecord.get(taskId).recordAttempt(nowMs);
+        } else {
+            taskIdToBackoffRecord.put(taskId, new BackoffRecord(nowMs));
+        }
+    }
+
+    public static class BackoffRecord {
+        private long attempts;
+        private long lastAttemptMs;
+        private static final ExponentialBackoff EXPONENTIAL_BACKOFF = new ExponentialBackoff(1, 2, 10000, 0.5);
+
+
+        public BackoffRecord(final long nowMs) {
+            this.attempts = 1;
+            this.lastAttemptMs = nowMs;
+        }
+
+        public void recordAttempt(final long nowMs) {
+            this.attempts++;
+            this.lastAttemptMs = nowMs;
+        }
+
+        public boolean canAttempt(final long nowMs) {
+            return  nowMs - lastAttemptMs >= EXPONENTIAL_BACKOFF.backoff(attempts);
+        }
     }
 }
